requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
starlette-cors
psycopg2-binary